Apply review cleanups: registry ordering, O(n) extraction, dedup
Registry (converter.cpp): the lenient DMP parser (accepts any buffer
>= 7 bytes) now sits last so the strict sniffers are reachable in
hint-less auto-detection, with the ordering rule documented once at the
list head. Making that rule true also required a signature check in the
RYM2612 parser, which previously default-filled a patch from any
non-XML input; it now requires a <PARAM id=" entry (real presets carry
~73). parse() no longer re-runs a failed hinted parser in the fallback
loop, and the Vgm entry's display name mentions VGZ.

VGM extraction (vgm.cpp): snapshot dedup switches from a linear rescan
per key-on to a hash set over the packed 48-byte channel state, which
also makes the prev_/prev_valid_ shadow copies redundant (removed);
grouping is a single pass keeping a running loudest representative per
group instead of two quadratic loops. The 0x52/0x53 and 0x68 command
special cases fold into the generic operand-length path — 0x68 now
gets the same truncation warning as every other command — and gunzip
returns the miniz heap buffer directly instead of copying it into a
vector. Output verified byte-identical on a real VGZ.

Also: drop noise casts in vgi/eif serialize to match tfi style, reuse
patches_equal/patches_strict_equal in the VGI/EIF roundtrip tests, add
the TFI spec URL to tfi.hpp, and list OPM among multi-patch formats in
the README.

+++ src/eif.cpp
@@ -81,21 +81,17 @@ ParseResult parse(const uint8_t *data, size_t size,
 SerializeResult serialize(const Patch &patch) {
   std::vector<uint8_t> data(kFileSize, 0);
 
-  data[0] = static_cast<uint8_t>((patch.algorithm & 0x07) |
-                                 ((patch.feedback & 0x07) << 3));
+  data[0] = (patch.algorithm & 0x07) | ((patch.feedback & 0x07) << 3);
 
   for (int op = 0; op < 4; ++op) {
     const auto &o = patch.operators[op];
 
-    data[1 + op] = static_cast<uint8_t>((o.ml & 0x0F) | ((o.dt & 0x07) << 4));
+    data[1 + op] = (o.ml & 0x0F) | ((o.dt & 0x07) << 4);
     data[5 + op] = std::min<uint8_t>(o.tl, 127);
-    data[9 + op] = static_cast<uint8_t>(std::min<uint8_t>(o.ar, 31) |
-                                        ((o.ks & 0x03) << 6));
-    data[13 + op] = static_cast<uint8_t>(std::min<uint8_t>(o.dr, 31) |
-                                         (o.am ? 0x80 : 0x00));
+    data[9 + op] = std::min<uint8_t>(o.ar, 31) | ((o.ks & 0x03) << 6);
+    data[13 + op] = std::min<uint8_t>(o.dr, 31) | (o.am ? 0x80 : 0x00);
     data[17 + op] = std::min<uint8_t>(o.sr, 31);
-    data[21 + op] = static_cast<uint8_t>(std::min<uint8_t>(o.rr, 15) |
-                                         ((o.sl & 0x0F) << 4));
+    data[21 + op] = std::min<uint8_t>(o.rr, 15) | ((o.sl & 0x0F) << 4);
     data[25 + op] = (o.ssg_enable ? 0x08 : 0x00) | (o.ssg & 0x07);
   }
